A lot of changes
Working on a new version of the hood survey and Javascript.

I was having trouble with the other one, so I built up a new one and
will not implement the CSS until I get the JS working.

Currently the script runs through but does not calculate neighborhood
like it should. I will be debugging this for awhile.